Change default sweep interval to 24 hours
moveFunds runs for a single user on command !tip balance so there is no need to schedule often.  Move funds also runs for a user on any send tip command.  The once a day will catch users who do not run any commands. Note: moveFunds is only if there is a balance in the zen address.